fix JS: JavaScpipt Long calculation

+++ kottage/data/indexeddb/src/jsMain/kotlin/io/github/irgaly/kottage/data/indexeddb/schema/entity/Item_event.kt
@@ -2,8 +2,8 @@ package io.github.irgaly.kottage.data.indexeddb.schema.entity
 
 external interface Item_event {
   var id: String
-  var created_at: Long
-  var expire_at: Long?
+  var created_at: Double
+  var expire_at: Double?
   var item_type: String
   var item_key: String
   var item_list_id: String?
