+++ kottage/data/indexeddb/src/jsMain/kotlin/io/github/irgaly/kottage/data/indexeddb/schema/entity/Item_list.kt
@@ -7,7 +7,7 @@ external interface Item_list {
   var item_key: String?
   var previous_id: String?
   var next_id: String?
-  var expire_at: Long?
+  var expire_at: Double?
   var user_info: String?
   var user_previous_key: String?
   var user_current_key: String?
