+++ kottage/data/indexeddb/src/jsMain/kotlin/io/github/irgaly/kottage/data/indexeddb/schema/entity/Item_stats.kt
@@ -2,6 +2,6 @@ package io.github.irgaly.kottage.data.indexeddb.schema.entity
 
 external interface Item_stats {
   var item_type: String
-  var count: Long
-  var event_count: Long
+  var count: Double
+  var event_count: Double
 }
